bot-backend/requirements.txt
azure-identity==1.22.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
Flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3